Basic_DataValidation

+++ api/Controllers/CommentController.cs
@@ -21,37 +21,70 @@ namespace api.Controllers
             _stockRepo = stockRepository;
         }
         [HttpGet]
-        public async Task<IActionResult> GetAll(){
+        public async Task<IActionResult> GetAll()
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             var comments = await _commentRepo.GetAllAsync();
+
             var commentsModel = comments.Select(comment => comment.ToDto());
+
             return Ok(commentsModel);
         }
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetById([FromRoute] int id){
+        [HttpGet("{id:int}")]
+        public async Task<IActionResult> GetById([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.GetByIdAsync(id);
-            if(comment == null) return NotFound();
+
+            if (comment == null)
+                return NotFound();
+
             return Ok(comment.ToDto());
         }
 
-        [HttpPost("{stockId}")]
-        public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto model){
-            if(!await _stockRepo.CheckStockExistance(stockId)) return BadRequest("Stock not found!");
+        [HttpPost("{stockId:int}")]
+        public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto model)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            if (!await _stockRepo.CheckStockExistance(stockId))
+                return BadRequest("Stock not found!");
+
             var comment = await _commentRepo.CreateAsync(model.ToEntityFromCreate(stockId));
-            return CreatedAtAction(nameof(GetById), new {id = comment.Id}, comment.ToDto());
+
+            return CreatedAtAction(nameof(GetById), new { id = comment.Id }, comment.ToDto());
         }
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto model){
-            // if(!await _stockRepo.CheckStockExistance(stockId)) return BadRequest("Stock not found!");
+        [HttpPut("{id:int}")]
+        public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto model)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.UpdateAsync(id, model.ToEntityFromUpdate());
-            if(comment == null) return NotFound("Comment not found!");
+
+            if (comment == null)
+                return NotFound("Comment not found!");
+
             return Ok(comment.ToDto());
         }
 
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> Delete([FromRoute] int id){
+        [HttpDelete("{id:int}")]
+        public async Task<IActionResult> Delete([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.DeleteAsync(id);
-            if(comment == null) return NotFound("Comment not found!");
+
+            if (comment == null)
+                return NotFound("Comment not found!");
+
             return Ok(comment.ToDto());
         }
     }
