+++ api/Controllers/StockController.cs
@@ -21,41 +21,68 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stocks = await _stockRepo.GetAllAsync();
+
             var stocksDto = stocks.Select(s => s.ToDto());
+
             return Ok(stocksDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stock = await _stockRepo.GetByIdAsync(id);
-            if (stock == null) return NotFound();
+
+            if (stock == null)
+                return NotFound();
+
             return Ok(stock.ToDto());
         }
 
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto model)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stock = await _stockRepo.CreateAsync(model.ToEntity());
+
             return CreatedAtAction(nameof(GetById), new { id = stock.Id }, stock.ToDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto model)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stock = await _stockRepo.UpdateAsync(id, model);
-            if (stock == null) return NotFound();
+
+            if (stock == null)
+                return NotFound();
+
             return Ok(stock.ToDto());
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stock = await _stockRepo.DeleteAsync(id);
-            if (stock == null) return NotFound();
+
+            if (stock == null)
+                return NotFound();
+
             return NoContent();
         }
     }
